Fixing port detection to also work on Windows
os.path.exists() raises an error when checking the existence of a serial
port under windows. Thus plottool is unuseable under Windows.

With checking the actual result of the serial.Serial() (aka opening the port)
we can check if the port could be opened and thus exists - even under windows.

This looks more like the pythonic way: Let it fail and maybe handle the error.

Signed-off-by: Chrissi^ <chris@tinyhost.de>